#69 refactoring: moving mapping implementation to the gateway; tests are not completely refactored yet
